Remove search + workspaces routes (extracted to separate branch)
Both routes transitively call libphonenumber-js, which breaks under
front-api's dev runtime (tsx wraps the CJS metadata JSON require in
\`{ default }\`). Until the proper fix lands (front-api ESM migration),
these two routes live on \`migrate-poke-search-workspaces\` alongside the
CJS/ESM mitigation.

Removed from this branch:
- Hono search + workspaces routes and their tests
- \`listWorkspacesForPoke\` lib extraction (+ \`getPlanCodeSortPriority\`
  helper that was only used by it)
- libphonenumber-js try/catch mitigation in \`lib/poke/search.ts\` and
  the workspaces lib (gone with the lib file)
- Migration markers on the Next search + workspaces handlers (those
  Next handlers remain canonical until the routes move back in)
