+++ front/pages/api/poke/search.ts
@@ -1,5 +1,4 @@
 /** @ignoreswagger */
-// @migration-status: MIGRATED_TO_HONO
 import { withSessionAuthenticationForPoke } from "@app/lib/api/auth_wrappers";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
